Finished sending a text message Fixed custom Lib Signal where the public keys have not been correctly padded, this caused the message to fail on sending. TODO: clear old chains it is currently keeping a copy instead of increasing the count

+++ WhatsSocket/Core/Models/Sessions/IndexInfo.cs
@@ -11,7 +11,7 @@ namespace WhatsSocket.Core.Models.Sessions
         public int BaseKeyType { get; set; }
 
         [JsonProperty("closed")]
-        public int Closed { get; set; }
+        public long Closed { get; set; }
 
         [JsonProperty("used")]
         public long Used { get; set; }
